cli start
